# docs/cmdb-api-spec.yaml
openapi: 3.0.3
info:
  title: CMDB API
  description: Chelas Movies Database API
  contact:
    name: André Monteiro
    email: a37316@alunos.isel.pt
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:1337
    description: Local server
...
paths:
...
  /api/groups:
    get:
      tags:
        - Groups
      summary: List all groups belonging to the user
      description: List all groups belonging to the user
      operationId: getGroups
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: All groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'
    post:
      tags:
        - Groups
      summary: Create group
      description: Create group
      operationId: createGroup
      security:
        - bearerAuth: [ ]
      requestBody:
        description: Group to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewGroupCreated'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'
...
  /api/users:
    post:
      tags:
        - Users
      summary: Create new user
      description: Adds a user to the system
      operationId: addUser
      requestBody:
        description: New User
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewUserCreated'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Movie:
      required:
        - id
        - title
        - year
        - runtimeMins
        - imdbRating
        - imageUrl
        - directors
        - writers
        - actors
      type: object
      properties:
        id:
          type: string
          description: Movie ID
          example: tt0111161
        title:
          type: string
          description: Movie title
          example: The Shawshank Redemption
        year:
          type: integer
          description: Movie release year
          example: 1994
        imdbRating:
          type: number
          description: Movie IMDB rating
          example: 9.3
        runtimeMins:
          type: integer
          minimum: 0
          description: Movie runtime
          example: 142
        imageUrl:
          type: string
          description: Movie image URL
          example: https://image.tmdb.org/t/p/w500/9O7gLzmreU0nGkIB6K3BsJbzvNv.jpg
        directors:
          type: string
          description: Movie director(s)
          example: Frank Darabont
        writers:
          type: string
          description: Movie writer(s)
          example: Stephen King
        actors:
          type: array
          description: Movie actors list
          items:
            $ref: '#/components/schemas/Actor'
...
    Group:
      required:
        - id
        - name
        - description
        - movies
        - totalDuration
      type: object
      properties:
        id:
          type: string
          description: Group ID
          example: 1
        name:
          type: string
          description: Group name
          example: My group
        description:
          type: string
          description: Group description
          example: My favorite movies
        movies:
          type: array
          description: Favorite movies list
          items:
            $ref: '#/components/schemas/Movie'
        totalDuration:
          type: integer
          minimum: 0
          description: Total duration in minutes of group movies
          example: 142
    NewGroup:
      required:
        - name
        - description
      type: object
      properties:
        name:
          type: string
          description: Group name
          example: My group
        description:
          type: string
          description: Group description
          example: My favorite movies
    NewGroupCreated:
      required:
        - status
        - group
      type: object
      properties:
        status:
          type: string
          description: Status
          example: New group created
        group:
          $ref: '#/components/schemas/Group'
    NewUser:
      required:
        - username
        - email
        - password
        - passwordConfirm
      type: object
      properties:
        username:
          type: string
          description: User name
          example: andre
        email:
          type: string
          description: email
          example: example@email.com
        password:
          type: string
          description: password
          example: password
        passwordConfirm:
          type: string
          description: passwordConfirm
          example: password
    NewUserCreated:
      required:
        - status
        - userId
        - username
        - email
        - token
      type: object
      properties:
        status:
          type: string
          description: status
          example: New user created
        userId:
          type: integer
          description: User ID
          example: 123456789
        username:
          type: string
          description: User name
          example: andre
        email:
          type: string
          description: E-mail address
          example: andre@example.com
        token:
          type: string
          description: User token
          format: uuid
          example: a3d50b64-e401-4c57-a698-809fa2b85af2
    Actor:
      required:
        - name
        - imageUrl
      type: object
      properties:
        name:
          type: string
          description: Actor name
          example: Tim Robbins
        imageUrl:
          type: string
          description: Actor image URL
          example: https://image.tmdb.org/t/p/w500/9O7gLzmreU0nGkIB6K3BsJbzvNv.jpg
...
    InvalidRequest:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
...
    NotAuthorized:
      type: object
      properties:
        error:
          type: string
          example: Not Authorized